requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database (compatible with Python 3.13)
sqlalchemy==2.0.35
alembic==1.13.1

# Redis
redis==5.0.1

# Email
aiosmtplib==3.0.1
dkimpy==1.1.5
dnspython==2.4.2
email-validator==2.1.0

# Async
aiohttp==3.9.1

# Cryptography
cryptography==42.0.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
uvicorn[standard]==0.24.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1